Created a spline-follow movement mode.
Created a custom mode Character Movement component that implements a spline-follow movement mode as its only custom mode. It stores a reference to a spline and follows that spline to completion, using MaxWalkSpeed for velocity.

The component is paired with an interface that allows one to:
- Set the spline to follow
- Initiate spline-follow
- Pause spline-follow

Both the component and the owning character (invader in this case) implement the interface.

Also created a minimal AI task to start the spline. In addition, spawn gate will pick a random path to follow on spawn.

Additional modifications:
- Created goblin controller and behavior tree to demo spline movement
- Modified goblin and orc to use goblint behavior tree
- Added basic run animation to goblin to test correct velocity updates