Implemented melee tower.
The melee tower is the base class for the Militia Barracks. Because it has simple behaviors, they are implemented directly in the tower class instead of in a separate behavior tree.

The melee tower first caches a starting rally point (once navigation is nailed down, we will switch to just using that to pick a random navigable point). Then, spawn three soldiers, assigning each one an index. On spawn, set the home location for the soldier, which it should then run to of its own accord.

When a soldier dies, they are respawned on a fixed timer.

Rally behavior and soldier AI is not yet implemented.