Implement Rally tower action.
This requires us to rethink our player input handling. When just dealing with abilities, it is fine to simply check whether an ability is selected or not. When introducing input handling for tower actions (like Rally), we need an all-encompassing enum for input state.

Accordingly, introduced EInputState, with various options for free input and targeting modes for ability and tower action.

- Added local enum EInputState with variants Free, TargetingAbility, and TargetingRally
- Added input state property to player controller
- Removed unused TestAbility()

For Rallying itself:

- Added getter for TowerBerth tower
- Created RallyCoordinator interface to support testing and setting of rally point
- Implemented RallyCoordinator interface to set new rally points
- Update Home and bRallying blackboard keys for soldiers on rally
- Created minimal task to unset bRallying flag
- Created minimal task to disengage when rallying