Created targeting zone for abilities and movement.
Targeting is different from navigability, in that navigable terrain is anything that can be moved across (which is basically the whole map) while targeting terrain is limited to area that can be targeted by an ability or a tower action like rallying.

The targeting area has been implemented to mirror the spline paths, with a bit of lateral padding. The targeting area has the custom collision channel "Targeting", which is reserved expressly for this purpose.